// src/interfaces/database/product.interface.ts
export interface Product {
    id: number;
    name: string | null;
    second_name: string | null;
    description: string | null;
    text_filter: string;
    model: string | null;
    factory_code: string | null;
    barcode: string | null;
    technical_specifications: string | null;
    item_type_id: string | null;
    internal_id: string | null;
    item_code: string | null;
    date_of_due: string | null;
    account_id: string | null;
    item_code_gs1: string | null;
    unit_type_id: string | null;
    meter: number | null;
    currency_type_id: string | null;
    sale_unit_price: number;
    purchase_has_igv: number;
    has_igv: number;
    subject_to_detraction: number;
    purchase_unit_price: number;
    has_isc: number;
    restrict_sale_cpe: number;
    exchange_points: number;
    quantity_of_points: number;
    commission_amount: number | null;
    line: string | null;
    commission_type: string | null;
    amount_plastic_bag_taxes: number;
    system_isc_type_id: string | null;
    percentage_isc: number;
    suggested_price: number;
    purchase_has_isc: number;
    purchase_system_isc_type_id: string | null;
    purchase_percentage_isc: number;
    sale_affectation_igv_type_id: string | null;
    purchase_affectation_igv_type_id: string | null;
    calculate_quantity: number;
    sale_unit_price_set: number | null;
    is_set: number;
    favorite: number;
    category_id: number | null;
    brand_id: number | null;
    image: string | null;
    image_medium: string | null;
    image_small: string | null;
    stock: number;
    stock_min: number;
    has_plastic_bag_taxes: number;
    lot_code: string | null;
    lots_enabled: number;
    series_enabled: number;
    percentage_of_profit: number;
    has_perception: number;
    percentage_perception: number | null;
    attributes: any[]; // o define la estructura si lo sabes
    active: number;
    frequent: number | null;
    web_platform_id: number | null;
    created_at: string;
    updated_at: string;
    warehouse_id: number | null;
    status: number;
    apply_store: number;
    descuento_store: number | null;
    apply_restaurant: number;
    cod_digemid: string | null;
    sanitary: string | null;
    is_for_production: number;
    info_link: string | null;
    has_sizes: number;
    is_food_dealer: number;
    id_cupones: number | null;
    genero: string | null;
    price_item_007: number;
    video_url: string | null;
}